requirements_lock.txt
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
accelerate==0.25.0
datasets==2.16.1
tokenizers==0.15.0
safetensors==0.4.1
huggingface-hub==0.20.2
wandb==0.16.2
tensorboard==2.15.1
numpy==1.26.3
pandas==2.1.4
pyarrow==14.0.2
tqdm==4.66.1
pyyaml==6.0.1
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
mypy==1.8.0
